CI: Download and build ebtree ourselves
Previously we used bintray, however the service has now shut down.
Checkout the repository and build instead (but cache the result).

This is a copy of commit e895d9d6fc10f7f8a8829c197b2c66d310e2156f in Ocean,
and was also applied to the other repositories.